[REF] Update Neighbour

+++ PathFindingProject/PathFindingProject/Vector2.cpp
@@ -12,7 +12,7 @@ Vector2 Vector2::operator-(const Vector2& vector2) const
 	return Vector2(x - vector2.x, y - vector2.y);
 }
 
-int Vector2::SqrDistance()
+float Vector2::SqrDistance()
 {
 	return std::sqrt(std::pow(x,2) + std::pow(y,2));
 }
